feat()!: Named Export
Using a factory that returns a class in TypeScript makes it very difficult (if not impossible) to emit proper type definitions as TypeScript has a hard time determining the types of the internal class.

The solution used in the the `typescript` branch works, however at the cost of proper JSDocs in production.

Alternetively A mixing pattern can be used but at the cost of private/protected methods and properties. I believe this to be too high of a cost.

After reviewing what other libraries are doing such as `connect-typeorm`, I have to the conclusion that import `Store` directly from `express-session` is the best play. While this does tie does directly to `express-session`, we already have them as a peer dependency so I believe this move makes sense.

`express-session` is by far the highest used session middleware these days. At least one that would require a store.

If  we decide to support other libraries in the future we can move `express-session` to a optional dependency, move all of it's imports to dynamic imports then make our own abstract class of `Store` that we can extend from to maintain type safety without the the vender lock in.

BREAKING CHANGE: Removes the default export factory in favor of a named export that imports `Store` directly from `express-session`

Signed-off-by: William Sedlacek <wsedlacekc@gmail.com>