fix: work-around fix for issue 88
Concurrent invocations of touch() or set() for the same session id can result in problems. Practically, this may happen when a browser concurrently loads multiple resources for the same site. This PR dis-allows concurrent calls to touch() or set() for the same session id by default, but provides options to over-ride this behavior.